🩹 disentangle `mqt-core-dd` and `mqt-core-algo`
