+++ test/algorithms/eval_dynamic_circuits.cpp
@@ -3,7 +3,6 @@
 #include "algorithms/QFT.hpp"
 #include "algorithms/QPE.hpp"
 #include "circuit_optimizer/CircuitOptimizer.hpp"
-#include "dd/Benchmark.hpp"
 #include "dd/DDDefinitions.hpp"
 #include "dd/Operations.hpp"
 #include "dd/Package.hpp"
@@ -177,10 +176,8 @@ TEST_P(DynamicCircuitEvalExactQPE, UnitaryTransformation) {
 TEST_P(DynamicCircuitEvalExactQPE, ProbabilityExtraction) {
   // generate DD of QPE circuit via simulation
   const auto start = std::chrono::steady_clock::now();
-  auto exp = dd::benchmarkSimulate(*qpe);
+  auto e = simulate(qpe.get(), dd->makeZeroState(qpe->getNqubits()), *dd);
   const auto simulationEnd = std::chrono::steady_clock::now();
-  auto e = exp->sim;
-  dd = std::move(exp->dd);
 
   // extract measurement probabilities from IQPE simulations
   dd::SparsePVec probs{};
@@ -390,10 +387,8 @@ TEST_P(DynamicCircuitEvalInexactQPE, ProbabilityExtraction) {
   std::cout << "---- extraction done ----\n";
 
   // generate DD of QPE circuit via simulation
-  auto exp = dd::benchmarkSimulate(*qpe);
+  auto e = simulate(qpe.get(), dd->makeZeroState(qpe->getNqubits()), *dd);
   const auto simulationEnd = std::chrono::steady_clock::now();
-  auto e = exp->sim;
-  dd = std::move(exp->dd);
   std::cout << "---- sim done ----\n";
 
   // compare outcomes
@@ -539,10 +534,8 @@ TEST_P(DynamicCircuitEvalBV, UnitaryTransformation) {
 TEST_P(DynamicCircuitEvalBV, ProbabilityExtraction) {
   // generate DD of QPE circuit via simulation
   const auto start = std::chrono::steady_clock::now();
-  auto exp = dd::benchmarkSimulate(*bv);
+  auto e = simulate(bv.get(), dd->makeZeroState(bv->getNqubits()), *dd);
   const auto simulationEnd = std::chrono::steady_clock::now();
-  auto e = exp->sim;
-  dd = std::move(exp->dd);
 
   // extract measurement probabilities from IQPE simulations
   dd::SparsePVec probs{};
@@ -687,12 +680,10 @@ TEST_P(DynamicCircuitEvalQFT, UnitaryTransformation) {
 TEST_P(DynamicCircuitEvalQFT, ProbabilityExtraction) {
   // generate DD of QPE circuit via simulation
   const auto start = std::chrono::steady_clock::now();
-  auto exp = dd::benchmarkSimulate(*qft);
+  auto e = simulate(qft.get(), dd->makeZeroState(qft->getNqubits()), *dd);
   const auto simulationEnd = std::chrono::steady_clock::now();
   const auto simulation =
       std::chrono::duration<double>(simulationEnd - start).count();
-  auto e = exp->sim;
-  dd = std::move(exp->dd);
 
   std::stringstream ss{};
   // extract measurement probabilities from IQPE simulations
