+++ test/algorithms/test_bernsteinvazirani.cpp
@@ -1,7 +1,6 @@
 #include "Definitions.hpp"
 #include "algorithms/BernsteinVazirani.hpp"
 #include "circuit_optimizer/CircuitOptimizer.hpp"
-#include "dd/Benchmark.hpp"
 #include "dd/Package.hpp"
 #include "dd/Simulation.hpp"
 
@@ -43,8 +42,10 @@ TEST_P(BernsteinVazirani, FunctionTest) {
   qc.printStatistics(std::cout);
 
   // simulate the circuit
+  auto dd = std::make_unique<dd::Package<>>(qc.getNqubits());
   const std::size_t shots = 1024;
-  auto measurements = dd::benchmarkSimulateWithShots(qc, shots);
+  auto measurements =
+      simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd, shots);
 
   for (const auto& [state, count] : measurements) {
     std::cout << state << ": " << count << "\n";
@@ -63,8 +64,10 @@ TEST_P(BernsteinVazirani, FunctionTestDynamic) {
   qc.printStatistics(std::cout);
 
   // simulate the circuit
+  auto dd = std::make_unique<dd::Package<>>(qc.getNqubits());
   const std::size_t shots = 1024;
-  auto measurements = dd::benchmarkSimulateWithShots(qc, shots);
+  auto measurements =
+      simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd, shots);
 
   for (const auto& [state, count] : measurements) {
     std::cout << state << ": " << count << "\n";
@@ -80,8 +83,10 @@ TEST_F(BernsteinVazirani, LargeCircuit) {
   qc.printStatistics(std::cout);
 
   // simulate the circuit
+  auto dd = std::make_unique<dd::Package<>>(qc.getNqubits());
   const std::size_t shots = 1024;
-  auto measurements = dd::benchmarkSimulateWithShots(qc, shots);
+  auto measurements =
+      simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd, shots);
 
   for (const auto& [state, count] : measurements) {
     std::cout << state << ": " << count << "\n";
@@ -97,8 +102,10 @@ TEST_F(BernsteinVazirani, DynamicCircuit) {
   qc.printStatistics(std::cout);
 
   // simulate the circuit
+  auto dd = std::make_unique<dd::Package<>>(qc.getNqubits());
   const std::size_t shots = 1024;
-  auto measurements = dd::benchmarkSimulateWithShots(qc, shots);
+  auto measurements =
+      simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd, shots);
 
   for (const auto& [state, count] : measurements) {
     std::cout << state << ": " << count << "\n";
