+++ test/algorithms/test_qft.cpp
@@ -1,9 +1,9 @@
 #include "algorithms/QFT.hpp"
-#include "dd/Benchmark.hpp"
 #include "dd/DDDefinitions.hpp"
 #include "dd/FunctionalityConstruction.hpp"
 #include "dd/Package.hpp"
 #include "dd/RealNumber.hpp"
+#include "dd/Simulation.hpp"
 
 #include <cassert>
 #include <cmath>
@@ -19,11 +19,16 @@ class QFT : public testing::TestWithParam<std::size_t> {
 protected:
   void TearDown() override {}
 
-  void SetUp() override { nqubits = GetParam(); }
+  void SetUp() override {
+    nqubits = GetParam();
+    dd = std::make_unique<dd::Package<>>(nqubits);
+  }
 
   std::size_t nqubits = 0;
+  std::unique_ptr<dd::Package<>> dd;
   std::unique_ptr<qc::QFT> qc;
-  std::unique_ptr<dd::Experiment> exp;
+  qc::VectorDD sim{};
+  qc::MatrixDD func{};
 };
 
 /// Findings from the QFT Benchmarks:
@@ -37,7 +42,7 @@ protected:
 /// The accuracy of double floating points allows for a minimal CN::TOLERANCE
 /// value of 10e-15
 ///	Utilizing more qubits requires the use of fp=long double
-constexpr std::size_t QFT_MAX_QUBITS = 20U;
+constexpr std::size_t QFT_MAX_QUBITS = 17U;
 
 static const size_t INITIAL_COMPLEX_COUNT = 1;
 
@@ -61,12 +66,8 @@ TEST_P(QFT, Functionality) {
   ASSERT_NO_THROW({ qc = std::make_unique<qc::QFT>(nqubits, false); });
   // there should be no error building the functionality
 
-  exp = dd::benchmarkFunctionalityConstruction(*qc);
-  auto* expFunc =
-      dynamic_cast<dd::FunctionalityConstructionExperiment*>(exp.get());
-  assert(expFunc != nullptr);
-  auto func = expFunc->func;
-  auto dd = std::move(exp->dd);
+  // there should be no error building the functionality
+  ASSERT_NO_THROW({ func = buildFunctionality(qc.get(), *dd); });
 
   qc->printStatistics(std::cout);
   // QFT DD should consist of 2^n nodes
@@ -111,12 +112,8 @@ TEST_P(QFT, FunctionalityRecursive) {
   // there should be no error constructing the circuit
   ASSERT_NO_THROW({ qc = std::make_unique<qc::QFT>(nqubits, false); });
 
-  exp = dd::benchmarkFunctionalityConstruction(*qc, true);
-  auto* expFunc =
-      dynamic_cast<dd::FunctionalityConstructionExperiment*>(exp.get());
-  assert(expFunc != nullptr);
-  auto func = expFunc->func;
-  auto dd = std::move(exp->dd);
+  // there should be no error building the functionality
+  ASSERT_NO_THROW({ func = buildFunctionalityRecursive(qc.get(), *dd); });
 
   qc->printStatistics(std::cout);
   // QFT DD should consist of 2^n nodes
@@ -161,11 +158,11 @@ TEST_P(QFT, Simulation) {
   // there should be no error constructing the circuit
   ASSERT_NO_THROW({ qc = std::make_unique<qc::QFT>(nqubits, false); });
 
-  exp = dd::benchmarkSimulate(*qc);
-  auto* expSim = dynamic_cast<dd::SimulationExperiment*>(exp.get());
-  auto sim = expSim->sim;
-  auto dd = std::move(exp->dd);
-
+  // there should be no error simulating the circuit
+  ASSERT_NO_THROW({
+    auto in = dd->makeZeroState(nqubits);
+    sim = simulate(qc.get(), in, *dd);
+  });
   qc->printStatistics(std::cout);
 
   // QFT DD |0...0> sim should consist of n nodes
@@ -199,15 +196,11 @@ TEST_P(QFT, FunctionalityRecursiveEquality) {
   ASSERT_NO_THROW({ qc = std::make_unique<qc::QFT>(nqubits, false); });
 
   // there should be no error building the functionality recursively
-  exp = dd::benchmarkFunctionalityConstruction(*qc);
-  auto* expFunc =
-      dynamic_cast<dd::FunctionalityConstructionExperiment*>(exp.get());
-  assert(expFunc != nullptr);
-  auto func = expFunc->func;
-  auto dd = std::move(exp->dd);
+  ASSERT_NO_THROW({ func = buildFunctionalityRecursive(qc.get(), *dd); });
 
   // there should be no error building the functionality regularly
-  auto funcRec = buildFunctionalityRecursive(qc.get(), *dd);
+  qc::MatrixDD funcRec{};
+  ASSERT_NO_THROW({ funcRec = buildFunctionality(qc.get(), *dd); });
 
   ASSERT_EQ(func, funcRec);
   dd->decRef(funcRec);
@@ -226,7 +219,8 @@ TEST_P(QFT, DynamicSimulation) {
 
   // simulate the circuit
   std::size_t shots = 8192U;
-  auto measurements = dd::benchmarkSimulateWithShots(*qc, shots);
+  auto measurements =
+      simulate(qc.get(), dd->makeZeroState(qc->getNqubits()), *dd, shots);
   for (const auto& [state, count] : measurements) {
     std::cout << state << ": " << count << "\n";
   }
